requirements.txt
azure-storage-blob==12.19.1
fastapi==0.110.0
pydantic==2.6.4
pyodbc==5.1.0
uvicorn==0.29.0